requirements.txt
aiohappyeyeballs==2.6.1 ; python_version >= "3.12" and python_version < "4"
aiohttp==3.11.18 ; python_version >= "3.12" and python_version < "4"
aiosignal==1.3.2 ; python_version >= "3.12" and python_version < "4"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4"
anyio==4.9.0 ; python_version >= "3.12" and python_version < "4"
asgiref==3.8.1 ; python_version >= "3.12" and python_version < "4"
astroid==3.3.9 ; python_version >= "3.12" and python_version < "4"
attrs==25.3.0 ; python_version >= "3.12" and python_version < "4"
audioop-lts==0.2.1 ; python_version >= "3.13" and python_version < "4"
black==25.1.0 ; python_version >= "3.12" and python_version < "4"
boto3==1.38.2 ; python_version >= "3.12" and python_version < "4"
botocore==1.38.2 ; python_version >= "3.12" and python_version < "4"
certifi==2025.1.31 ; python_version >= "3.12" and python_version < "4"
charset-normalizer==3.4.1 ; python_version >= "3.12" and python_version < "4"
click==8.1.8 ; python_version >= "3.12" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4" and (platform_system == "Windows" or sys_platform == "win32")
deprecated==1.2.18 ; python_version >= "3.12" and python_version < "4"
dill==0.4.0 ; python_version >= "3.12" and python_version < "4"
discord-py==2.5.2 ; python_version >= "3.12" and python_version < "4"
fastapi==0.115.12 ; python_version >= "3.12" and python_version < "4"
frozenlist==1.6.0 ; python_version >= "3.12" and python_version < "4"
googleapis-common-protos==1.70.0 ; python_version >= "3.12" and python_version < "4"
grpcio==1.71.0 ; python_version >= "3.12" and python_version < "4"
h11==0.16.0 ; python_version >= "3.12" and python_version < "4"
idna==3.10 ; python_version >= "3.12" and python_version < "4"
importlib-metadata==8.6.1 ; python_version >= "3.12" and python_version < "4"
isort==6.0.1 ; python_version >= "3.12" and python_version < "4"
jmespath==1.0.1 ; python_version >= "3.12" and python_version < "4"
limits==5.1.0 ; python_version >= "3.12" and python_version < "4"
mccabe==0.7.0 ; python_version >= "3.12" and python_version < "4"
multidict==6.4.3 ; python_version >= "3.12" and python_version < "4"
mypy-extensions==1.1.0 ; python_version >= "3.12" and python_version < "4"
mypy==1.15.0 ; python_version >= "3.12" and python_version < "4"
opentelemetry-api==1.32.1 ; python_version >= "3.12" and python_version < "4"
opentelemetry-distro==0.53b1 ; python_version >= "3.12" and python_version < "4"
opentelemetry-exporter-otlp-proto-common==1.32.1 ; python_version >= "3.12" and python_version < "4"
opentelemetry-exporter-otlp-proto-grpc==1.32.1 ; python_version >= "3.12" and python_version < "4"
opentelemetry-exporter-otlp-proto-http==1.32.1 ; python_version >= "3.12" and python_version < "4"
opentelemetry-exporter-otlp==1.32.1 ; python_version >= "3.12" and python_version < "4"
opentelemetry-instrumentation-asgi==0.53b1 ; python_version >= "3.12" and python_version < "4"
opentelemetry-instrumentation-fastapi==0.53b1 ; python_version >= "3.12" and python_version < "4"
opentelemetry-instrumentation==0.53b1 ; python_version >= "3.12" and python_version < "4"
opentelemetry-proto==1.32.1 ; python_version >= "3.12" and python_version < "4"
opentelemetry-sdk==1.32.1 ; python_version >= "3.12" and python_version < "4"
opentelemetry-semantic-conventions==0.53b1 ; python_version >= "3.12" and python_version < "4"
opentelemetry-util-http==0.53b1 ; python_version >= "3.12" and python_version < "4"
packaging==25.0 ; python_version >= "3.12" and python_version < "4"
pathspec==0.12.1 ; python_version >= "3.12" and python_version < "4"
platformdirs==4.3.7 ; python_version >= "3.12" and python_version < "4"
propcache==0.3.1 ; python_version >= "3.12" and python_version < "4"
protobuf==5.29.4 ; python_version >= "3.12" and python_version < "4"
pydantic-core==2.33.1 ; python_version >= "3.12" and python_version < "4"
pydantic==2.11.3 ; python_version >= "3.12" and python_version < "4"
pylint==3.3.6 ; python_version >= "3.12" and python_version < "4"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4"
s3transfer==0.12.0 ; python_version >= "3.12" and python_version < "4"
sentry-sdk==2.27.0 ; python_version >= "3.12" and python_version < "4"
six==1.17.0 ; python_version >= "3.12" and python_version < "4"
slowapi==0.1.9 ; python_version >= "3.12" and python_version < "4"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4"
starlette==0.46.2 ; python_version >= "3.12" and python_version < "4"
structlog==25.2.0 ; python_version >= "3.12" and python_version < "4"
tomlkit==0.13.2 ; python_version >= "3.12" and python_version < "4"
typing-extensions==4.13.2 ; python_version >= "3.12" and python_version < "4"
typing-inspection==0.4.0 ; python_version >= "3.12" and python_version < "4"
urllib3==2.4.0 ; python_version >= "3.12" and python_version < "4"
uvicorn==0.34.2 ; python_version >= "3.12" and python_version < "4"
wrapt==1.17.2 ; python_version >= "3.12" and python_version < "4"
yarl==1.20.0 ; python_version >= "3.12" and python_version < "4"
zipp==3.21.0 ; python_version >= "3.12" and python_version < "4"
